scripts/requirements.txt
click==7.1.2
requests==2.25.1
wheel==0.36.2
asyncclick==7.1.2.3
PyNaCl==1.5.0
kiba-core[queues, requester, web3]==0.5.2.dev10
